[ Spring Boot 2 ] 1. Book.java      - PK int -> Long 변경 ========== REST - CRUD =========== 2. BookRepository.java      1) JPA 제공 메소드      2) 쿼리 메소드      3) 사용자 정의 메소드 (@Query()) 3. BookService.java 4. BookRestController.java      - REST Controller
======================================
5. BookController.java
     - 라우팅 역할을 수행하는 Controller
6. list.html
     - view

+++ src/main/java/com/example/demo/entity/Book.java
@@ -9,7 +9,7 @@ import javax.persistence.*;
 public class Book {
     @Id // PK
     @GeneratedValue(strategy = GenerationType.IDENTITY) // Auto-increament
-    private int id;     // 일련번호
+    private Long id;     // 일련번호
 
     @Column(length = 50, nullable = false)
     private String title;
